Se termina el flujo 'feliz' del proyecto. TODO: 1 - presicion del resultado no es exacta. 2 - Abrir archivo que cargue todo,hasta instancias y te lleve a la primer tab - 3 eliminar botones de siguiente y atras. 4 - ver manejo del viewport que se rompe cuando crece. 5 -arreglar flechas que tengan un punto de inflección. 6 - deshacer, y modificar barra de herramientas segun tab. 7 - la instancia se rompe cuando vas y volves .. verlo con tab

+++ src/main/java/ar/unsl/qualisys/componentes/nodos/QsNodo.java
@@ -21,7 +21,7 @@ public class QsNodo extends JPanel {
     //2664554567
     private String padreID;
     private double valorResultado; 
-    private float ponderacion;
+    private double ponderacion;
 
     /**
      * Constructor que van a usar las variables/que para el caso son constantes por eso no hay "set"
@@ -61,11 +61,11 @@ public class QsNodo extends JPanel {
         this.valorResultado = valorResultado;
     }
 
-    public float getPonderacion() {
+    public double getPonderacion() {
         return ponderacion;
     }
 
-    public void setPonderacion(float ponderacion) {
+    public void setPonderacion(double ponderacion) {
         this.ponderacion = ponderacion;
     }
